// src/app/core/generated-service/model/login-response.ts
/**
 * Silicone API
 * Web API for Silicone
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserDto } from './user-dto';
import { MenuDto } from './menu-dto';


export interface LoginResponse { 
    token?: string | null;
    refreshToken?: string | null;
    sessionId?: string;
    userInfo?: UserDto;
    menus?: Array<MenuDto> | null;
}